Refactor AzkabanUploadTask and move session Authentication and other helper methods by creating new AzkabanHelper class
