+++ CONTRIBUTORS.md
@@ -74,3 +74,5 @@ The following were contributed by Ragesh Rajagopalan. Thanks, Ragesh!
 * `LIHADOOP-20985 azkabanUpload throwing 500 error`
 * `DI-584 Login prompt missing when uploading to Azkaban`
 
+The following were contributed by Pranay Hasan Yerra. Thanks, Pranay!
+* 'LIHADOOP-21658 Refactor AzkabanUploadTask in Hadoop Plugin by creating AzkabanHelper class'
