+++ gradle.properties
@@ -1,2 +1,2 @@
 org.gradle.daemon=true
-version=0.9.9
+version=0.9.10
